create securite in Activite

+++ api/Service/activiteService.php
@@ -1,21 +1,28 @@
 <?php
 
 include_once './Repository/activiteRepository.php';
+include_once './index.php';
 
 class activiteService {
 
     /*
      *  Récupère tous les Activite
     */
-    public function getAllActivite() {
-        $ActiviteRepository = new ActiviteRepository();
-        return $ActiviteRepository->getAllActivite();
+    public function getAllActivite($apikey) {
+        $userRole = getRoleFromApiKey($apikey);
+        if ($userRole==1 || $userRole==2 || $userRole==4) {
+            $ActiviteRepository = new ActiviteRepository();
+            return $ActiviteRepository->getAllActivite();
+        }else{
+            exit_with_message("Vous n'avais pas accès a cette commande");
+        }
     }
 
     /*
      *  Récupère un Activite par son id
     */
-    public function getActiviteById($id) {
+    public function getActiviteById($id,$apikey) {
+        $userRole = getRoleFromApiKey($apikey);
         $ActiviteRepository = new ActiviteRepository();
         return $ActiviteRepository->getActiviteById($id);
     }
@@ -23,9 +30,14 @@ class activiteService {
     /*
      *  Créer un Activite
     */
-    public function createActivite(ActiviteModels $Activite) {
-        $ActiviteRepository = new ActiviteRepository();
-        return $ActiviteRepository->createActivite($Activite);
+    public function createActivite(ActiviteModels $Activite, $apikey) {
+        $userRole = getRoleFromApiKey($apikey);
+        if ($userRole==1 || $userRole==2 || $userRole==4) {
+            $ActiviteRepository = new ActiviteRepository();
+            return $ActiviteRepository->createActivite($Activite);
+        }else{
+            exit_with_message("Vous n'avais pas accès a cette commande");
+        }
     }
 
     /*
@@ -44,9 +56,14 @@ class activiteService {
     /*
      *  Supprimer un Activite
     */
-    public function deleteActivite($id) {
-        $ActiviteRepository = new ActiviteRepository();
-        return $ActiviteRepository->deleteActivite($id);
+    public function deleteActivite($id,$apikey) {
+        $userRole = getRoleFromApiKey($apikey);
+        if ($userRole==1 || $userRole==2) {
+            $ActiviteRepository = new ActiviteRepository();
+            return $ActiviteRepository->deleteActivite($id);
+        }else{
+            exit_with_message("Vous n'avais pas accès a cette commande");
+        }
     }
 }
 ?>
